requirements.txt
langchain==0.2.*
langchain-community==0.2.*
langchain-openai==0.1.8
faiss-cpu==1.8.0
tiktoken==0.7.0
pypdf==4.2.0
python-dotenv==1.0.1
